+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Upgrade to Webpack 4 @sneridagh
 * Renew login token when almost expired @robgietema
 * Added actions reducers @robgietema
+* Add error reporting with Sentry support on client (default ErrorBoundary), server and Redux middleware @sneridagh
 
 ## 0.4.0 (2017-05-03)